redesigned UI to fit changes
removed the scrollable chips at the the top as I discussed with the team that that will be needed in a different screen. Added the search area button to update the list of toilets and render them onto the screen based on what region the user moved the map to. I also added a footer that will navigate to the home screen, list and user screens but currently is not attached to anything. the map type button not works and switches between satellite and standard maps